% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqldf.R
\docType{package}
\name{DBIsqldf-package}
\alias{DBIsqldf-package}
\title{DBIsqldf package overview}
\description{
Provides an easy way to perform SQL selects on R data frames.
}
\details{
The package contains a single function \code{\link[=sqldf]{sqldf()}} whose help file
contains more information and examples.

This package is a fork from the \href{https://github.com/ggrothendieck/sqldf}{original sqldf-package}.
The primary differences are:
\itemize{
\item DBI interface only. This increases compatibility with database drivers, but at the expense
of database-specific convenience features.
\item DBI native parameterized query support.
\item More conservative defaults for writing tables - data frames are imported as temporary tables
and should be committed explicitly is they need to be persisted.
\item No legacy (transitive) dependencies.
}
}
